fix(secret-store): into_iter should output String, avoid a breaking change
